Restructured
Restructured the repo and removed some projects. A lot of them built off each other and contained the same code, so I added the final projects whenever I could. This repo is now a lot smaller and should help give someone a better idea of what I'm capable of, but they can also see my growth as a programmer.